Removed unnecessary cast.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/io/trunk@610516 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/io/input/CharSequenceReader.java
@@ -79,7 +79,7 @@ public class CharSequenceReader extends Reader implements Serializable {
         if (idx >= charSequence.length()) {
             return -1;
         } else {
-            return (int)charSequence.charAt(idx++);
+            return charSequence.charAt(idx++);
         }
     }
